+++ src/utils/constants.js
@@ -1,0 +1,12 @@
+export const profile = {
+  firstname: 'Fabien',
+  lastname: 'Gréard',
+  email: 'grd.fabien@gmail.com',
+  mobile: '+33 6 24 36 09 35',
+};
+
+export const socials = {
+  linkedin: 'https://www.linkedin.com/in/fabiengreard/',
+  github: 'https://github.com/FabienGreard',
+  stackOverflow: 'https://stackoverflow.com/users/6668441/fabien-greard',
+};