+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.6](https://github.com/FabienGreard/fabiengreard/compare/v0.0.5...v0.0.6) (2021-02-15)
+
+
+### Features
+
+* lightouse-ci ([75415bf](https://github.com/FabienGreard/fabiengreard/commit/75415bff3c73c65d4bcc546e6b82446e57ba3c0e))
+* pwa ([bb76548](https://github.com/FabienGreard/fabiengreard/commit/bb765480e78ce4ce8ea5fa19bc5b202069cf8270))
+* sw, contact ([da3a7cd](https://github.com/FabienGreard/fabiengreard/commit/da3a7cd4faa26182fbcb1e4714a0c493327b4b23))
+
+
+### Bug Fixes
+
+* fallback color ([ef39398](https://github.com/FabienGreard/fabiengreard/commit/ef39398daae996088ef0d2b4e016078bb52440b7))
+* lighthouse-ci ([5290b83](https://github.com/FabienGreard/fabiengreard/commit/5290b83ff765811d44d0ea5f42901f30d5b36246))
+* lighthouse-ci ([a471546](https://github.com/FabienGreard/fabiengreard/commit/a471546ea4a8bf7313f1f25c0c045c920b4e279a))
+* link title ([2e85bc1](https://github.com/FabienGreard/fabiengreard/commit/2e85bc1ae1dfffc57cd4deb1effbc637c9ead563))
+* node 12 requirement for lighthouse-ci ([59ce3d9](https://github.com/FabienGreard/fabiengreard/commit/59ce3d9d6449817a90430893d62d6fb624538f70))
+* og-image ([67a54d7](https://github.com/FabienGreard/fabiengreard/commit/67a54d70964180f443ca7cbfa7dc6ba93dee5d39))
+* output path revert ([660bad1](https://github.com/FabienGreard/fabiengreard/commit/660bad1b7129a55e30c16ef91165329d8d78d6aa))
+* overflow of width on mobile device ([65c28f5](https://github.com/FabienGreard/fabiengreard/commit/65c28f5f4956694a68c0ca550b30cc6641e258d8))
+* prefetch link ?? ([057b47d](https://github.com/FabienGreard/fabiengreard/commit/057b47ded17df7c055fb0aed27fef9cb17c379c8))
+* preload ([8653f16](https://github.com/FabienGreard/fabiengreard/commit/8653f1688d3f8bf751fd0b0519417e4237a3b9d1))
+* preload clean ([5c13ed0](https://github.com/FabienGreard/fabiengreard/commit/5c13ed0aa8459aed8d368d8aba4835e2ac8cacd9))
+* preload css ([68fcd40](https://github.com/FabienGreard/fabiengreard/commit/68fcd40f6c1fce1456deaa49d9eb33ac780287e7))
+* remove cssminifi ([2a1eb6a](https://github.com/FabienGreard/fabiengreard/commit/2a1eb6ac70f351bcc46de2393bc0aae730b01e94))
+* revert obsolet webpack plugin ([9a53911](https://github.com/FabienGreard/fabiengreard/commit/9a539112ba7ed1d88291a7b8d662c44b20b3009f))
+* script name ([d08e51b](https://github.com/FabienGreard/fabiengreard/commit/d08e51b1a634c0448ec9c1bb46867779427faa15))
+* sw in static ([f09893c](https://github.com/FabienGreard/fabiengreard/commit/f09893ca8324fa4294412053d547d78d96684235))
+
 ### [0.0.5](https://github.com/FabienGreard/fabiengreard/compare/v0.0.4...v0.0.5) (2021-02-13)